requirements.txt
python-pptx
pptx_tables==0.1.0